refactor : adjust step2 review & split package with step2 & step3
