+++ src/main/java/step3/InputView.java
@@ -1,3 +1,5 @@
+package step3;
+
 import java.util.Scanner;
 
 public class InputView {
